bridge: use buffered output when writing message
This fixes ipc connections that try to read 14 bytes
at once to get the header, and fail to handle
cases when a shorter amount of bytes have been written.

By using a buffered writer, we can guarantee that the header
will be always complete when the first write syscall happens.